Reformating dates on previews
